+++ kartothek/io_components/read.py
@@ -6,10 +6,11 @@ import pandas as pd
 from kartothek.core.factory import DatasetFactory
 from kartothek.core.index import ExplicitSecondaryIndex
 from kartothek.io_components.metapartition import MetaPartition
-from kartothek.io_components.utils import _make_callable
+from kartothek.io_components.utils import _make_callable, normalize_args
 from kartothek.serialization import check_predicates, columns_in_predicates
 
 
+@normalize_args
 def dispatch_metapartitions_from_factory(
     dataset_factory,
     label_filter=None,
@@ -75,7 +76,7 @@ def dispatch_metapartitions_from_factory(
 
         # Group the resulting MetaParitions by partition keys or a subset of those keys
         merged_partitions = base_df.groupby(
-            by=list(dispatch_by), sort=False, as_index=False
+            by=list(dispatch_by), sort=True, as_index=False
         )
         for group_name, group in merged_partitions:
             if not isinstance(group_name, tuple):