+++ tests/io/dask/dataframe/test_read.py
@@ -1,9 +1,14 @@
 import pickle
 from functools import partial
 
+import dask
+import dask.dataframe as dd
+import numpy as np
 import pandas as pd
 import pytest
+from dask.dataframe.utils import assert_eq as assert_dask_eq
 from pandas import testing as pdt
+from pandas.testing import assert_frame_equal
 
 from kartothek.io.dask.dataframe import read_dataset_as_ddf
 from kartothek.io.eager import store_dataframes_as_dataset
@@ -104,3 +109,109 @@ def test_read_ddf_from_categorical_partition(store_factory):
     )
     df_actual = ddf.compute(scheduler="sync")
     pdt.assert_frame_equal(df, df_actual)
+
+
+@pytest.mark.parametrize("index_type", ["primary", "secondary"])
+def test_reconstruct_dask_index(store_factory, index_type, monkeypatch):
+    dataset_uuid = "dataset_uuid"
+    colA = "ColumnA"
+    colB = "ColumnB"
+    df1 = pd.DataFrame({colA: [1, 2], colB: ["x", "y"]})
+    df2 = pd.DataFrame({colA: [3, 4], colB: ["x", "y"]})
+    df_chunks = np.array_split(pd.concat([df1, df2]).reset_index(drop=True), 4)
+    df_delayed = [dask.delayed(c) for c in df_chunks]
+    ddf_expected = dd.from_delayed(df_delayed).set_index(
+        colA, divisions=[1, 2, 3, 4, 4]
+    )
+    ddf_expected_simple = dd.from_pandas(
+        pd.concat([df1, df2]), npartitions=2
+    ).set_index(colA)
+
+    if index_type == "secondary":
+        secondary_indices = colA
+        partition_on = None
+    else:
+        secondary_indices = None
+        partition_on = colA
+
+    store_dataframes_as_dataset(
+        store=store_factory,
+        dataset_uuid=dataset_uuid,
+        dfs=[df1, df2],
+        secondary_indices=secondary_indices,
+        partition_on=partition_on,
+    )
+
+    # Make sure we're not shuffling anything
+    monkeypatch.delattr(
+        dask.dataframe.shuffle, dask.dataframe.shuffle.set_index.__name__
+    )
+    ddf = read_dataset_as_ddf(
+        dataset_uuid=dataset_uuid,
+        store=store_factory,
+        table="table",
+        dask_index_on=colA,
+    )
+
+    assert ddf_expected.npartitions == 4
+    assert len(ddf_expected.divisions) == 5
+    assert ddf_expected.divisions == (1, 2, 3, 4, 4)
+    assert ddf.index.name == colA
+
+    assert ddf.npartitions == 4
+    assert len(ddf.divisions) == 5
+    assert ddf.divisions == (1, 2, 3, 4, 4)
+
+    assert_dask_eq(ddf_expected, ddf)
+    assert_dask_eq(ddf_expected_simple, ddf, check_divisions=False)
+
+    assert_frame_equal(ddf_expected.compute(), ddf.compute())
+    assert_frame_equal(ddf_expected_simple.compute(), ddf.compute())
+
+
+def test_reconstruct_dask_index_sorting(store_factory, monkeypatch):
+
+    # Make sure we're not shuffling anything
+    monkeypatch.delattr(
+        dask.dataframe.shuffle, dask.dataframe.shuffle.set_index.__name__
+    )
+    dataset_uuid = "dataset_uuid"
+    colA = "ColumnA"
+    colB = "ColumnB"
+
+    df = pd.DataFrame(
+        {colA: np.random.randint(high=100000, low=-100000, size=(50,)), colB: 0}
+    )
+    store_dataframes_as_dataset(
+        store=store_factory, dataset_uuid=dataset_uuid, dfs=[df], partition_on=colA
+    )
+    ddf = read_dataset_as_ddf(
+        dataset_uuid=dataset_uuid,
+        store=store_factory,
+        table="table",
+        dask_index_on=colA,
+    )
+
+    assert all(
+        ddf.map_partitions(lambda df: df.index.min()).compute().values
+        == ddf.divisions[:-1]
+    )
+
+
+def test_reconstruct_dask_index_raise_no_index(store_factory):
+    dataset_uuid = "dataset_uuid"
+    colA = "ColumnA"
+    df1 = pd.DataFrame({colA: [1, 2]})
+    store_dataframes_as_dataset(
+        store=store_factory, dataset_uuid=dataset_uuid, dfs=[df1]
+    )
+    with pytest.raises(
+        RuntimeError,
+        match=r"Requested index: \['ColumnA'\] but available index columns: \[\]",
+    ):
+        read_dataset_as_ddf(
+            dataset_uuid=dataset_uuid,
+            store=store_factory,
+            table="table",
+            dask_index_on=colA,
+        )